refactoring Home css
