--- CSS/card.sass
@@ -1,28 +1,0 @@
-@import "./colors.module.sass"
-
-.card 
-  margin: 1rem
-  padding: 1.5rem
-  text-align: left
-  color: inherit
-  text-decoration: none
-  border: 1px solid colors.$border-light
-  border-radius: 10px
-  transition: color 0.15s ease, border-color 0.15s ease
-  max-width: 300px
-
-  &:hover,
-  &:focus,
-  &:active 
-    color: colors.$secondary
-    border-color: colors.$secondary
-
-  & h2 
-    margin: 0 0 1rem 0
-    font-size: 1.5rem
-
-
-  & p 
-    margin: 0
-    font-size: 1.25rem
-    line-height: 1.5
