+++ pages/index.tsx
@@ -2,48 +2,43 @@ import Head from 'next/head'
 import Image from 'next/image'
 import Link from "next/link"
 import React from 'react'
-import BalanceTable from '../components/BalanceTable'
-import styles from '../CSS/Home.module.css'
-import css from '../CSS/styles.module.sass'
-import pageStyles from '../CSS/page.module.sass'
+import BalanceTable from '../components/tables/BalanceTable'
+import homeStyles from '../CSS/Home.module.sass'
 import { DefaultPage } from '../components/DefaultPage'
 
 export default function Home() {
 
   return <>
-        <Head>
-        <title>Portfolio</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-  <DefaultPage 
-    title="Your Portfolio!" 
-    description="">
-<div className={styles.container}>
+    <Head>
+      <title>Portfolio</title>
+      <meta name="description" content="Generated by create next app" />
+      <link rel="icon" href="/favicon.ico" />
+    </Head>
+    <DefaultPage title="Your Portfolio!">
+    <div className={homeStyles.container}>
 
-      <main className={styles.main}>
-        <div></div>
+      <main className={homeStyles.main}>
 
         <BalanceTable />
 
-        <div className={styles.grid}>
+        <div className={homeStyles.grid}>
           
           <Link href="/currencies">
-            <a className={css.card}> 
+            <a className={homeStyles.card}> 
               <h2>Currencies &rarr;</h2>
               <p>Fiat and Crypto currencies.</p>
             </a>
           </Link>
 
           <Link href="/companies">
-            <a className={css.card}> 
+            <a className={homeStyles.card}> 
               <h2>Companies &rarr;</h2>
               <p>Banks, Exchanges and other companies where you can store funds.</p>
             </a>
           </Link>
 
           <Link href="/settings">
-            <a className={css.card}> 
+            <a className={homeStyles.card}> 
               <h2>Settings &rarr;</h2>
               <p>Application and user settings.</p>
             </a>
